+++ src/curve/arithmetic.rs
@@ -1,4 +1,4 @@
-use alloc::{format, string::String, vec, vec::Vec};
+use alloc::{boxed::Box, format, string::String, vec, vec::Vec};
 use core::ops::{Add, Mul, Neg, Rem, Sub};
 
 use digest::XofReader;
@@ -25,6 +25,7 @@ use ::{ecdsa::SigningKey, elliptic_curve::SecretKey};
 use crate::{
     params::SchemeParams,
     tools::{hashing::Chain, BoxedRng, Secret},
+    uint::BoxedEncoding,
 };
 
 pub(crate) fn chain_curve<Crv, Chn>(digest: Chn) -> Chn
@@ -198,6 +199,19 @@ where
         Self(<P::Curve as CurveArithmetic>::Scalar::conditional_select(
             &a.0, &b.0, choice,
         ))
+    }
+}
+
+impl<P> BoxedEncoding for Scalar<P>
+where
+    P: SchemeParams,
+{
+    fn to_be_bytes(&self) -> Box<[u8]> {
+        (*self).to_be_bytes().as_ref().into()
+    }
+
+    fn try_from_be_bytes(bytes: &[u8]) -> Result<Self, String> {
+        Self::try_from_be_bytes(bytes)
     }
 }